+++ lib/components/inline.js
@@ -1,4 +1,5 @@
 const { SourceFile, Buffer } = require('../file')
+const { docify } = require('./wrappers')
 
 /**
  * Inline declarations of types can come in different flavours.
@@ -10,24 +11,24 @@ class InlineDeclarationResolver {
 
     /**
      * @param {string} name
-     * @param {import('../compile').TypeResolveInfo} type
+     * @param {import('./resolver').TypeResolveInfo} type
      * @param {import('../file').Buffer} buffer
      * @param {string} statementEnd
-     * @private
+     * @protected
      * @abstract
      */
     printInlineType(name, type, buffer, statementEnd) { /* abstract */ }
 
     /**
      * Attempts to resolve a type that could reference another type.
-     * @param {any} val 
-     * @param {import('../compile').TypeResolveInfo} into @see Visitor.resolveType
+     * @param {any} items 
+     * @param {import('./resolver').TypeResolveInfo} into @see Visitor.resolveType
      * @param {SourceFile} file temporary file to resolve dummy types into.
      * @public
      */
     resolveInlineDeclaration(items, into, relativeTo) {
         const dummy = new SourceFile(relativeTo.path.asDirectory())
-        dummy.classes.currentIndet = relativeTo.classes.currentIndent
+        dummy.classes.currentIndent = relativeTo.classes.currentIndent
         dummy.classes.add('{')
         dummy.classes.indent()
 
@@ -36,7 +37,7 @@ class InlineDeclarationResolver {
             // in inline definitions, we sometimes have to resolve first
             // FIXME: does this tie in with how we sometimes end up with resolved typed in resolveType()?
             const se = (typeof subelement === 'string')
-                ? this.visitor._resolveTypeName(subelement)
+                ? this.visitor.resolver.resolveTypeName(subelement)
                 : subelement
             into.structuredType[subname] = this.visitor.visitElement(subname, se, dummy)
         }
@@ -55,17 +56,17 @@ class InlineDeclarationResolver {
     /**
      * Visits a single element in an entity.
      * @param {string} name name of the element
-     * @param {import('../compile').CSN} element CSN data belonging to the the element.
+     * @param {import('./resolver').CSN} element CSN data belonging to the the element.
      * @param {SourceFile} file the namespace file the surrounding entity is being printed into.
      * @param {Buffer} [buffer] buffer to add the definition to. If no buffer is passed, the passed file's class buffer is used instead.
      * @public
      */
     visitElement(name, element, file, buffer = file.classes) {
         this.depth++
-        for (const d of this.visitor._docify(element.doc)) {
+        for (const d of docify(element.doc)) {
             buffer.add(d)
         }
-        const type = this.visitor.resolveAndRequire(element, file)
+        const type = this.visitor.resolver.resolveAndRequire(element, file)
         this.depth--
         if (this.depth === 0) {
             this.printInlineType(name, type, buffer)
@@ -77,13 +78,13 @@ class InlineDeclarationResolver {
      * Separator between value V and type T: `v : T`.
      * Depending on the visitor's setting, this is may be `?:` for optional
      * properties or `:` for required properties.
-     * @returns {'?'|':'}
+     * @returns {'?:'|':'}
      */
     getPropertyTypeSeparator() {
         return this.visitor.options.propertiesOptional ? '?:' : ':'
     }
 
-    /** @param {import('../compile').Visitor} visitor */
+    /** @param {import('../visitor').Visitor} visitor */
     constructor(visitor) {
         this.visitor = visitor
         // type resolution might recurse. This indicator is used to determine
@@ -110,7 +111,7 @@ class InlineDeclarationResolver {
      * @public
      * @abstract
      */
-    getTypeLookup(members) { /* abstract */ }
+    getTypeLookup(members) { /* abstract */ return '' }
 }
 
 /**
