+++ lib/file.js
@@ -137,6 +137,19 @@ class SourceFile extends File {
     }
 
     /**
+     * Adds a function definition in form of a arrow function to the file.
+     * @param {string} name name of the function
+     * @param {{relative: string | undefined, local: boolean, posix: boolean}} params list of parameters, passed as [name, type] pairs
+     * @param returns the return type of the function
+     */
+    addFunction(name, params, returns) {
+        // FIXME: use different buffers for buffers and actions, or at least rename buffer to the more general category "functions"?
+        this.actions.buffer.add("// function")
+        this.actions.buffer.add(`export declare const ${SourceFile.stringifyLambda(name, params, returns)};`)
+        this.actions.names.push(name)
+    }
+
+    /**
      * Adds an action definition in form of a arrow function to the file.
      * @param {string} name name of the action
      * @param {{relative: string | undefined, local: boolean, posix: boolean}} params list of parameters, passed as [name, type] pairs
@@ -428,6 +441,7 @@ class Path {
  * @type {SourceFile}
  */
 const baseDefinitions = new SourceFile('_')
+// FIXME: this should be a library someday
 baseDefinitions.addPreamble(`
 export namespace Association {
     export type to <T> = T;
