+++ lib/logging.js
@@ -16,15 +16,16 @@ class Logger {
         for (let i = 0; i < lvls.length - 1; i++) {
             // -1 to ignore NONE
             const level = lvls[i]
-            this[level.toLowerCase()] = function (message) {
-                this._log(level, message)
-            }.bind(this)
+            this[level.toLowerCase()] = function (message) { this._log(level, message) }.bind(this)
         }
     }
 
+    // only temporarily to disable those warnings...
+    //warning(s) {}; error(s) {}; info(s) {}; debug(s) {};
+
     /**
      * Add all log levels starting at level.
-     * @param {number} level level to start from.
+     * @param {number} baseLevel level to start from.
      */
     addFrom(baseLevel) {
         const vals = Object.values(Levels)
