+++ lib/util.js
@@ -61,9 +61,9 @@ const unlocalize = (name) => {
  * Attempts to derive the singular form of an English noun.
  * If '@singular' is passed as annotation, that is preferred.
  * @param {Annotations} dn annotations
- * @param {boolean} stripped if true, leading namespace will be stripped
+ * @param {boolean?} stripped if true, leading namespace will be stripped
  */
-const singular4 = (dn, stripped) => {
+const singular4 = (dn, stripped = false) => {
     let n = dn.name || dn
     if (stripped) {
         n = n.match(last)[0]
@@ -184,6 +184,50 @@ const parseCommandlineArgs = (argv, validFlags) => {
     }
 }
 
+/** 
+* Entities inherit their ancestors annotations:
+* https://cap.cloud.sap/docs/cds/cdl#annotation-propagation
+* This is a problem if we annotate @singular/ @plural to an entity A,
+* as we don't want all descendents B, C, ... to share the ancestor's
+* annotated inflexion
+* -> remove all such annotations that appear in a parent as well.
+* BUT: we can't just delete the attributes. Imagine three classes
+* A <- B <- C
+* where A contains a @singular annotation.
+* If we erase the annotation from B, C will still contain it and
+* can not detect that its own annotation was inherited without
+* travelling up the entire inheritance chain up to A.
+* So instead, we monkey patch and maintain a dictionary "erased"
+* when removing an annotation which we also check.
+*/
+function fixCSN(csn) {
+    const erase = (entity, parent, attr) => {
+        if (attr in entity) {
+            const ea = entity[attr]
+            if (parent[attr] === ea || (parent.erased && parent.erased[attr] === ea)) {
+                entity.erased ??= {}
+                entity.erased[attr] = ea
+                delete entity[attr]
+                this.logger.info(`Removing inherited attribute ${attr} from ${entity.name}.`)
+            }
+        }
+    }
+
+    for (const entity of Object.values(csn.definitions)) {
+        let i = 0
+        if (entity.includes) {
+            while (
+                (getSingularAnnotation(entity) || getPluralAnnotation(entity)) &&
+                i < entity.includes.length
+            ) {
+                const parent = this.csn.definitions[entity.includes[i]]
+                Object.values(annotations).flat().forEach(an => erase(entity, parent, an))
+                i++
+            }
+        }
+    }
+}
+
 module.exports = {
     annotations,
     getSingularAnnotation,
@@ -193,4 +237,5 @@ module.exports = {
     plural4,
     parseCommandlineArgs,
     deepMerge,
+    fixCSN
 }
