+++ test/unit/files/actions/elsewhere.cds
@@ -1,0 +1,3 @@
+namespace elsewhere;
+type ExternalType {}
+type ExternalType2 {}
