+++ test/unit/files/actions/model.cds
@@ -1,4 +1,7 @@
 namespace actions_test;
+using elsewhere.ExternalType from './elsewhere';
+using elsewhere.ExternalType2 from './elsewhere';
+using ExternalInRoot from './root';
 
 service S {
     entity E {}
@@ -6,7 +9,15 @@ service S {
             action f (x: String);
             function g (a: { x: Integer; y: Integer }, b: Integer) returns Integer;
             action h () returns { a: Integer; b: String };
+            action k () returns ExternalType;
+            action l () returns ExternalInRoot;
         }
 }
 
 action free (param: String) returns { a: Integer; b: String } ;
+action free2 () returns ExternalType;
+action free3 () returns ExternalInRoot;
+
+action free4 () returns {
+    foo: ExternalType2
+};
