+++ test/unit/files/actions/root.cds
@@ -1,0 +1,1 @@
+type ExternalInRoot {};
