+++ test/unit/files/arrayof/elsewhere.cds
@@ -1,0 +1,2 @@
+namespace elsewhere;
+type ExternalType {}
